surge_poc/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pyyaml==6.0.2
python-multipart==0.0.9
streamlit==1.39.0
requests==2.32.3